requirements.txt
aiohttp~=3.12.12
faiss-cpu~=1.11.0
cryptography~=45.0.4
numpy~=2.3.0
pandas~=2.3.0
pydantic-settings~=2.9.1
pydantic~=2.11.7
pytest~=8.4.0
python-json-logger~=3.3.0
py_clob_client~=0.23.0
requests~=2.32.4
sentence-transformers~=4.1.0
sortedcontainers~=2.4.0
spacy~=3.8.7
supabase~=2.15.3
torch~=2.7.1
typing_extensions~=4.12.2
websockets~=14.2
web3~=7.12.0